requirements.txt
aiohttp==3.9.5
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
elastic-transport==8.12.0
elasticsearch==8.12.1
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.6
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pydantic==2.6.3
pydantic_core==2.16.3
pyrate-limiter==2.10.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
ratelimiter==1.2.0.post0
requests==2.31.0
requests-ratelimiter==0.6.0
schedule==1.2.1
semver==3.0.2
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
